Add currency conversion functionality in car routes
- Updated car retrieval endpoints to support currency conversion based on user input.
- Introduced exception handling for invalid currency codes.
- Added new service methods for fetching cars and car details with currency conversion.

+++ backend/currency_converter/client.py
@@ -23,7 +23,7 @@ class CurrencyConverterClient:
         return self.client.service.getAvailableCurrencies()
 
     def convert(self, from_currency: str, to_currency: str, amount: Decimal) -> Decimal:
-        price_in_cent = amount * 100
+        price_in_cent = int(amount * 100)
         converted_price_in_cent = self.client.service.convert(from_currency, to_currency, price_in_cent)
         return Decimal(converted_price_in_cent) / Decimal('100')
